2024-5-14累积更新

+++ Libs/DBLib/MySQLRecordSet.cpp
@@ -274,7 +274,7 @@ bool CMySQLRecordSet::Close()
 	return true;
 }
 
-bool CMySQLRecordSet::SetBlobMaxProcessSize(UINT64 MaxSize)
+bool CMySQLRecordSet::SetBlobMaxProcessSize(size_t MaxSize)
 {
 	return false;
 }
